Enforce think time in PseudoServerSocket.
Before this change, the PseudoServerSocket plainly relied on the client
not trying to read from it before it gets a notification. However, the
BackTap implementation does in fact try to read more often - essentially
every time its token bucket is refilled. Therefore, BackTap wouldn't wait
for the processing time to be over, but read data early (and thus get a
certain "unfair" advantage over e.g. the vanilla Tor implementation).